fix(unitcore): compute ut_encode_clock in double to remove latent overflow
ut_encode_clock's int expression (hours*60 + minutes)*60 overflows int32
(signed-overflow UB) for nonsensical inputs; the binding term is the
outer *60, so the limit is hours ~ 596,523. Computing in double (the
return type) removes the overflow with no int64_t/long, and so no LLP64
long-is-32-bit wrinkle. Every integer up to 2^53 is exact in double, so
the result is bit-identical for all realistic inputs.

Confirmed under UBSan: the int form trips at 36000000 * 60 and returns
garbage; the double form is trip-free and returns the correct value. The
function stays a side-effect-free encoder -- guard still removed, not
re-coupled to ut_check_clock, no status set.

Tests (testDateTime.c): add test_ut_encode_clock_total_no_overflow --
bit-identical at realistic inputs, correct/finite past the old int
limit, NaN second still propagates.

Ref: PR148-review-response_stage1.md item 2 (A1)

+++ lib/unitcore.c
@@ -422,7 +422,17 @@ ut_encode_clock(
     int		minutes,
     double	seconds)
 {
-    return (hours*60 + minutes)*60 + seconds;
+    /*
+     * Compute in double (the return type) rather than int. The int form
+     * (hours*60 + minutes)*60 overflows int32 (signed-overflow UB) for
+     * nonsensical inputs -- the binding term is the outer *60, so the limit is
+     * hours ~ 596,523. Computing in double removes the overflow entirely with
+     * no int64_t/long (and so no LLP64 long-is-32-bit wrinkle). Every integer
+     * up to 2^53 is exact in double, so the result is bit-identical for all
+     * realistic inputs. This stays a side-effect-free encoder: no status is
+     * set (see the status convention in udunits2.h).
+     */
+    return ((double)hours*60 + minutes)*60 + seconds;
 }
 
 inline static int
